+++ src/main/java/com/brand/blockus/blocks/blockitems/LegacyBlockItem.java
@@ -6,6 +6,7 @@ import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.tooltip.TooltipType;
+import net.minecraft.screen.ScreenTexts;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Util;
@@ -13,9 +14,11 @@ import net.minecraft.util.Util;
 import java.util.List;
 
 public class LegacyBlockItem extends BlockItem {
+    private final String version;
 
-    public LegacyBlockItem(Block block, Settings settings) {
+    public LegacyBlockItem(Block block, Settings settings, String version) {
         super(block, settings);
+        this.version = version;
     }
 
     @Override
@@ -23,5 +26,8 @@ public class LegacyBlockItem extends BlockItem {
         super.appendTooltip(stack, context, tooltip, options);
         this.getBlock().appendTooltip(stack, context, tooltip, options);
         tooltip.add(Text.translatable(Util.createTranslationKey("blockitem", Blockus.id("legacy"))).formatted(Formatting.LIGHT_PURPLE));
+        tooltip.add(ScreenTexts.EMPTY);
+        tooltip.add(Text.literal(version).formatted(Formatting.ITALIC).formatted(Formatting.GRAY));
+
     }
 }