+++ src/main/java/com/brand/blockus/datagen/models/BlockusTextureKey.java
@@ -1,4 +1,4 @@
-package com.brand.blockus.data.models;
+package com.brand.blockus.datagen.models;
 
 import net.minecraft.data.client.TextureKey;