+++ changelog.md
@@ -4,12 +4,19 @@
   * Can be placed in 2 directions.
   * Connects to wall signs.
   * Crafted with 3 logs
+* Changed Small Logs recipe, now crafted with 4 logs in triangle form and give 4.
 * Added Copper Bricks
   * Oxidizes and can be waxed like every copper blocks
   * Crafted with 4 Cut Copper of the desired oxidation or waxed.
 * Added Mossy Dripstone Bricks
-* Changed Small Logs recipe, now crafted with 4 logs in triangle form and give 4.
 * Changed Polished Dripstone, Dripstone Bricks, Cracked Dripstone Bricks & Chiseled Dripstone textures
+* Changed Glowing Obsidian texture
+* Added both Legacy Roses (Java & PE).
+  * Unlike original version, Legacy Rose from PE give light blue dye.
+  * Like original version, Legacy Roses gives 2 dyes instead of one.
+  * Java one is crafted with 4 Poppies and gives 2, PE one is crafted with 2 Legacy Roses & gives 2.
+* Legacy Sapling recipe now gives 4 saplings instead of one.
+* Added the version on which the block/texture is based in the legacy blocks tooltip
 * Tweaked Cherry & Mangrove Small Logs textures.
 * Removed unused barrels translation strings
 * Fixed fireproof timber frames, wooden grates & lattices items.